[EFL][WK2] Let Ecore_Evas own the cursor object https://bugs.webkit.org/show_bug.cgi?id=100731
Patch by Christophe Dumez <christophe.dumez@intel.com> on 2012-10-30
Reviewed by Kenneth Rohde Christiansen.

Ecore_Evas takes care of calling evas_object_del() on
the cursor object when it is no longer used (e.g. when
it is replaced by another cursor object). Therefore,
we don't need to keep a RefPtr to the cursor Evas
Object as a data member of EwkViewImpl. We pass
ownership of the cursor object to Ecore_Evas.

* UIProcess/API/efl/EwkViewImpl.cpp:
(EwkViewImpl::setCursor):
* UIProcess/API/efl/EwkViewImpl.h:
(EwkViewImpl):

Identifier: 118764@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132880 268f45cc-cd09-0410-ab3c-d52691b4dbfc