[touchadjustment] touch-links-longpress tests passes incorrectly https://bugs.webkit.org/show_bug.cgi?id=96810
Patch by Rick Byers <rbyers@chromium.org> on 2012-10-30
Reviewed by Antonio Gomes.

Tools:

Add support for suppling width/height information for longpress
gesture (as for other gestures like tapDown) for the purposes
of touch adjustment.

* DumpRenderTree/chromium/TestRunner/src/EventSender.cpp:
(WebTestRunner):
(WebTestRunner::EventSender::gestureEvent):

LayoutTests:

Fix touch-links-longpress test to be consistent with touch-links-active.
Now it's actually touching where we intended - at the center of the
touch region identified instead of the top/left corner.  Also pass
the width/height of the touch to EventSender in order to enable touch
adjustment.

Also adds checks to verify that our touches are actually landing on
or outside the element we're targetting as desired.

* touchadjustment/resources/touchadjustment.js:
(findAbsoluteBounds): Update to support both inline and block nodes (clientHeight is 0 for inline nodes)
* touchadjustment/touch-links-active.html: Use updated findAbsoluteBounds
* touchadjustment/touch-links-longpress.html:

Identifier: 118813@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132929 268f45cc-cd09-0410-ab3c-d52691b4dbfc