Crash in WebProcess at com.apple.WebCore: WebCore::ResourceLoader::start <rdar://problem/12596761> and https://webkit.org/b/100792
Reviewed by Tim Hatcher.

In release builds we can sometimes end up with a null ResourceLoader due to a race condition
with the WebProcess telling the NetworkProcess to remove a resource load at the very same time
the NetworkProcess is about to start that load.

Until we can resolve that race condition an early return will be necessary.

* WebProcess/Network/WebResourceLoadScheduler.cpp:
(WebKit::WebResourceLoadScheduler::startResourceLoad):

Identifier: 118831@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132950 268f45cc-cd09-0410-ab3c-d52691b4dbfc