[BlackBerry] PurgeCredential should also purge the persisted credential store https://bugs.webkit.org/show_bug.cgi?id=100744
Reviewed by George Staikos.

When challenged by the server, we will check the credential in-memory and from the persistent storage.
So the invalid credentials should also be purged from the persistent credential storage; Otherwise,
it will be sent to the server next time been challenged, instead of prompting the user.

* platform/network/blackberry/NetworkJob.cpp:
(WebCore::NetworkJob::purgeCredentials):


Identifier: 118855@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@132976 268f45cc-cd09-0410-ab3c-d52691b4dbfc