[CSS Shaders] Validate types of built-in uniforms https://bugs.webkit.org/show_bug.cgi?id=98974
Patch by Max Vujovic <mvujovic@adobe.com> on 2012-10-31
Reviewed by Dean Jackson.

Source/WebCore:

Reject shaders which define built-in uniforms with the wrong type. For example, we reject a
shader with the GLSL code "uniform float u_textureSize;" because u_textureSize should be a
vec2, not a float.

Relevant Spec Section:
https://dvcs.w3.org/hg/FXTF/raw-file/tip/filters/index.html#shader-uniform-variables

Test: css3/filters/custom/invalid-custom-filter-uniform-types.html

* platform/graphics/filters/CustomFilterValidatedProgram.cpp:
(WebCore):
(WebCore::builtInUniformNameToTypeMap):
(WebCore::validateSymbols):

LayoutTests:

Add a test to verify that shaders which define built-in uniforms with the wrong type do not
execute. Add shaders for each built-in uniform in CSS Custom Filters.

Relevant Spec Section:
https://dvcs.w3.org/hg/FXTF/raw-file/tip/filters/index.html#shader-uniform-variables

* css3/filters/custom/invalid-custom-filter-uniform-types-expected.html: Added.
* css3/filters/custom/invalid-custom-filter-uniform-types.html: Added.
* css3/filters/resources/invalid-type-u-mesh-box.fs: Added.
* css3/filters/resources/invalid-type-u-mesh-size.fs: Added.
* css3/filters/resources/invalid-type-u-projection-matrix.fs: Added.
* css3/filters/resources/invalid-type-u-texture-size.fs: Added.
* css3/filters/resources/invalid-type-u-tile-size.fs: Added.
* css3/filters/resources/invalid-type-uniform-array.fs: Added.

Identifier: 118955@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133083 268f45cc-cd09-0410-ab3c-d52691b4dbfc