SVG as an image may recreate the renderer on zoom https://bugs.webkit.org/show_bug.cgi?id=99508
Reviewed by Abhishek Arya.

Source/WebCore:

The SVGImage code, when SVG is used in <img> tags, caches the renderer
at the start of the painting method and re-uses the pointer at the end
of the method. However, when the page is zoomed the renderer may be
detached mid-method, thus leaving a stray pointer. The fix is to
re-fetch the pointer after the zooms.

Test: svg/as-image/img-zoom-svg-stylesheet.html

* svg/graphics/SVGImage.cpp:
(WebCore::SVGImage::drawSVGToImageBuffer): Re-fetch the renderer after
the zoom operations.

LayoutTests:

Test and ref-test-result for zooming of SVG content when loaded as-image, and the svg is actually a css file.

* svg/as-image/img-zoom-svg-stylesheet-expected.html: Added.
* svg/as-image/img-zoom-svg-stylesheet.html: Added.
* svg/as-image/resources/svg-stylesheet-for-zoom.svg: Added.


Identifier: 119022@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133155 268f45cc-cd09-0410-ab3c-d52691b4dbfc