Unreviewed GTK gardening.
Skipping ENABLE_SUBPIXEL- and ENABLE_INPUT_TYPE_*-related testing
directories. These features are not enabled yet and most of the
tests fail.

Removing failure expectation for fast/files/workers/worker-apply-blob-url-to-xhr.html,
the issue was fixed in r132973.

Moving the skip entry for css3/device-adapt to the proper section.

* platform/gtk/TestExpectations:


Identifier: 119279@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133444 268f45cc-cd09-0410-ab3c-d52691b4dbfc