[Chromium] Enable the JavaScript i18n API for Android https://bugs.webkit.org/show_bug.cgi?id=101357
Reviewed by Adam Barth.

Source/WebCore:

Remove the conditional dependency on the v8-i18n API. The library is now
always available in a Chromium for Android checkout.

* WebCore.gyp/WebCore.gyp:

Source/WebKit/chromium:

Enable the JavaScript i18n API for Chromium on Android. This adds a
grand total of 18.7 KiB of data to the content shell apk.

* features.gypi:


Identifier: 119459@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133642 268f45cc-cd09-0410-ab3c-d52691b4dbfc