AX: Textfields don't get focus when navigated to from 'show all tabs' button https://bugs.webkit.org/show_bug.cgi?id=101369
Reviewed by Darin Adler.

When focus leaves a webpage, the webpage does not normally clear its focused node. When focus
returns (through tab, or mouse click), then focus will be cleared and restored.

However, when focus is returned through accessibility means, focus is not cleared first, which
means subsequent focus actions do nothing.

This fix will check if the focused node is the same as the node being focused, and if so, clear the focus first.

Making a layout test for this is not possible because it requires setting focus to something in the chrome, and then
moving focus back to the webpage.

* accessibility/AccessibilityRenderObject.cpp:
(WebCore::AccessibilityRenderObject::setFocused):


Identifier: 119582@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@133785 268f45cc-cd09-0410-ab3c-d52691b4dbfc